reafactor(ui): TMS-2519 rename tenant dialog
Renamed TenantCreateDialog to TenantrequestDialog
